# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The backend powering most things",
        "title": "web-api",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/ystv/web-api",
            "email": "computing@ystv.co.uk"
        },
        "license": {}
    },
    "paths": {
...
        "/v1/internal/creator/encodes/presets": {
            "get": {
                "description": "Lists all encode presets, these are groups of instructions (formats) for the encoder to create the video",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "creator-encodes"
                ],
                "summary": "List all encode presets",
                "operationId": "get-creator-encodes-presets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/encode.Preset"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "updates an preset",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "creator-encodes"
                ],
                "summary": "Update a preset",
                "operationId": "update-creator-encodes-preset",
                "parameters": [
                    {
                        "description": "Preset object",
                        "name": "quote",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/encode.Preset"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            },
            "post": {
                "description": "creates a new preset.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "creator-encodes"
                ],
                "summary": "New preset",
                "operationId": "new-creator-encodes-preset",
                "parameters": [
                    {
                        "description": "Preset object",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/encode.Preset"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Preset ID",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
...
        "/v1/internal/creator/playlists": {
            "get": {
                "description": "Lists all playlists, doesn't include videos inside.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "creator-playlists"
                ],
                "summary": "List all playlists",
                "operationId": "get-creator-playlists-all",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/playlist.Playlist"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update a playlist, video ID's required otherwise it will remove all videos.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "creator-playlists"
                ],
                "summary": "Update playlist",
                "operationId": "update-creator-playlist",
                "parameters": [
                    {
                        "description": "Playlist object",
                        "name": "quote",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/playlist.New"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            },
            "post": {
                "description": "creates a new playlist with optional video ID's.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "creator-playlists"
                ],
                "summary": "New playlist",
                "operationId": "new-creator-playlist",
                "parameters": [
                    {
                        "description": "Playlist object",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/playlist.Playlist"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Playlist ID",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
        "/v1/internal/creator/playlists/{playlistid}": {
            "get": {
                "description": "Get a playlist including it's children videos.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "creator-playlists"
                ],
                "summary": "Get playlist by ID",
                "operationId": "get-creator-playlist",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Playlist ID",
                        "name": "playlistid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/playlist.Playlist"
                        }
                    }
                }
            }
        },
...
        "/v1/internal/creator/series/{seriesid}": {
            "get": {
                "description": "Get a series including it's children videos.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "creator-series"
                ],
                "summary": "Get series by ID",
                "operationId": "get-creator-series",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Series ID",
                        "name": "seriesid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/series.Series"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "encode.Format": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "mimeType": {
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "watermarked": {
                    "type": "boolean"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "encode.Preset": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "formats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/encode.Format"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "playlist.New": {
            "type": "object",
            "properties": {
                "createdBy": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                },
                "videoIDs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "playlist.Playlist": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                },
                "videos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/video.Meta"
                    }
                }
            }
        },
...
        "series.Meta": {
            "type": "object",
            "properties": {
                "depth": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "series.Series": {
            "type": "object",
            "properties": {
                "childSeries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/series.Meta"
                    }
                },
                "depth": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "videos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/video.Meta"
                    }
                }
            }
        },
...
        "video.Meta": {
            "type": "object",
            "properties": {
                "broadcastDate": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "presetID": {
                    "type": "string"
                },
                "seriesID": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                },
                "userNickname": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
...
    }
}